fix: suggestiong for stg-5
