+++ Assets/Scenes/Level_1.unity
@@ -47178,7 +47178,8 @@ MonoBehaviour:
     - {fileID: 810101520}
   - enemyPrefab: {fileID: 1602361522707910691, guid: 30085eb2a544dda4cb35c9fb1e2703bb, type: 3}
     spawnPoint: {fileID: 263139674}
-    destinationPoints: []
+    destinationPoints:
+    - {fileID: 263139674}
   - enemyPrefab: {fileID: 1602361522707910691, guid: a5ebbff9a96d6c142af6fffc1cf1cef1, type: 3}
     spawnPoint: {fileID: 1825266251}
     destinationPoints:
