+++ Assets/Scripts/Enemy/Enemy.cs
@@ -3,17 +3,16 @@ using UnityEngine;
 
 public class Enemy : NetworkBehaviour {
     public Health health;
-    public NetworkObject networkObject;
 
-    void Awake() {
+    void Start() {
         if (health == null) { Debug.Log("Health not set to enemy!"); return; }
-        if (networkObject == null) { Debug.Log("Network object not set to enemy!"); return; }
+        if (NetworkObject == null) { Debug.Log("Network object not set to enemy!"); return; }
         health.onDie.AddListener(EnemyDestroyLocal);
     }
 
     public void EnemyDestroyLocal() {
         if (IsServer)
-            networkObject.Despawn();
+            NetworkObject.Despawn();
         else
             RequestDespawnServerRpc();
     }
@@ -21,6 +20,6 @@ public class Enemy : NetworkBehaviour {
     [ServerRpc(RequireOwnership = false)]
     private void RequestDespawnServerRpc() {
         if (!IsServer) return;
-        networkObject.Despawn(destroy: true);
+        NetworkObject.Despawn(destroy: true);
     }
 }
