+++ Assets/Scripts/Gameplay.asmdef
@@ -1,11 +1,18 @@
 {
-  "name": "Gameplay",
-  "autoReferenced": true,
-  "references": [
-    "Unity.Netcode.Runtime",
-    "Unity.Netcode",
-    "Unity.TextMeshPro",
-    "Unity.InputSystem",
-    "Health"
-  ]
+    "name": "Gameplay",
+    "rootNamespace": "",
+    "references": [
+        "Unity.Netcode.Runtime",
+        "Unity.TextMeshPro",
+        "Unity.InputSystem"
+    ],
+    "includePlatforms": [],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": [],
+    "versionDefines": [],
+    "noEngineReferences": false
 }
