+++ Assets/Scripts/Paralax.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bc478d8c6d7211546a8c79ec0131a20e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
