+++ Assets/Scripts/Paralax/CameraRotation.cs.meta
