+++ Assets/Scripts/Paralax/Prefab Manager.cs.meta
